Fixed GameTree and DCFR tests

+++ include/trainable/Trainable.h
@@ -16,20 +16,20 @@ class Trainable {
  public:
   virtual ~Trainable() = default;
 
-  virtual const std::vector<float>& GetCurrentStrategy() const = 0;
-  virtual const std::vector<float>& GetAverageStrategy() const = 0;
+  virtual const std::vector<double>& GetCurrentStrategy() const = 0;
+  virtual const std::vector<double>& GetAverageStrategy() const = 0;
 
   // *** REVERT: Use scalar weight for regret update ***
-  virtual void UpdateRegrets(const std::vector<float>& regrets, int iteration,
+  virtual void UpdateRegrets(const std::vector<double>& regrets, int iteration,
                              double reach_prob_opponent_chance_scalar) = 0; // Renamed for clarity
 
   // *** REVERT: Use scalar weight for average strategy update ***
   // *** FIX 6: Rename based on user feedback/common practice ***
-  virtual void AccumulateAverageStrategy(const std::vector<float>& current_strategy,
+  virtual void AccumulateAverageStrategy(const std::vector<double>& current_strategy,
                                          int iteration,
                                          double reach_prob_player_chance_scalar) = 0; // Renamed & scalar
 
-  virtual void SetEv(const std::vector<float>& evs) = 0;
+  virtual void SetEv(const std::vector<double>& evs) = 0;
   virtual json DumpStrategy(bool with_ev) const = 0;
   virtual json DumpEvs() const = 0;
   virtual void CopyStateFrom(const Trainable& other) = 0;
